map PlayerPointWithPlayer to PlayerInfo in server action
